# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.CreateAccount.createAccountRequest:
    properties:
      account_number:
        type: string
      amount:
        type: number
      color:
        type: string
      currency:
        type: string
      is_main_account:
        type: boolean
      issuer:
        type: string
      type:
        enum:
        - saving-account
        - credit-loan
        - goal-driven-saving
        type: string
    required:
    - account_number
    - currency
    - issuer
    - type
    type: object
...
  controllers.Deposit.depositRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
...
  controllers.Transfer.transferRequest:
    properties:
      amount:
        type: number
      from_account_id:
        type: string
      to_account_id:
        type: string
    required:
    - amount
    - from_account_id
    - to_account_id
    type: object
  controllers.UpdateAccount.updateAccountRequest:
    properties:
      account_number:
        type: string
      color:
        type: string
      currency:
        type: string
      issuer:
        type: string
      progress:
        maximum: 100
        minimum: 0
        type: integer
      type:
        enum:
        - saving-account
        - credit-loan
        - goal-driven-saving
        type: string
    type: object
...
  controllers.Withdraw.withdrawRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
  models.AccountFlag:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      deleted_at:
        description: for soft delete
        type: string
      flag_id:
        type: integer
      flag_type:
        description: system and user
        type: string
      flag_value:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - account_id
    - flag_type
    - flag_value
    - user_id
    type: object
  models.AccountWithDetails:
    properties:
      account_id:
        description: Account fields
        type: string
      account_number:
        type: string
      amount:
        description: AccountBalance fields
        type: number
      color:
        description: AccountDetail fields
        type: string
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      flags:
        description: AccountFlags
        items:
          $ref: '#/definitions/models.AccountFlag'
        type: array
      is_main_account:
        type: boolean
      issuer:
        type: string
      progress:
        type: integer
      type:
        description: saving-account, credit-loan, goal-driven-saving
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Banner:
    properties:
      banner_id:
        type: string
      created_at:
        type: string
      deleted_at:
        description: for soft delete
        type: string
      description:
        type: string
      image:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - banner_id
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: suthasinee.ler@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Developer Assignment API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Get all accounts for a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AccountWithDetails'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: Account details
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateAccount.createAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AccountWithDetails'
      security:
      - ApiKeyAuth: []
      summary: Create account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Get a single account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountWithDetails'
      security:
      - ApiKeyAuth: []
      summary: Get account
      tags:
      - accounts
    patch:
      consumes:
      - application/json
      description: Update an existing account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account details
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateAccount.updateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountWithDetails'
      security:
      - ApiKeyAuth: []
      summary: Update account
      tags:
      - accounts
  /accounts/{id}/deposit:
    post:
      consumes:
      - application/json
      description: Deposit money to an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Amount to deposit
        in: body
        name: amount
        required: true
        schema:
          $ref: '#/definitions/controllers.Deposit.depositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deposit money
      tags:
      - accounts
  /accounts/{id}/main:
    put:
      consumes:
      - application/json
      description: Set an account as the main account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Set main account
      tags:
      - accounts
  /accounts/{id}/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraw money from an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Amount to withdraw
        in: body
        name: amount
        required: true
        schema:
          $ref: '#/definitions/controllers.Withdraw.withdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Withdraw money
      tags:
      - accounts
  /accounts/transfer:
    post:
      consumes:
      - application/json
      description: Transfer money between accounts
      parameters:
      - description: Transfer details
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/controllers.Transfer.transferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Transfer money
      tags:
      - accounts
...
  /banners:
    get:
      description: List all banners for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Banner'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List banners
      tags:
      - Banners
  /banners/{id}:
    get:
      description: Get a banner by its ID
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
      security:
      - ApiKeyAuth: []
      summary: Get banner by ID
      tags:
      - Banners
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"